sepalate integration tests for treekem and enclave_key
