+++ scripts/test.sh
@@ -52,13 +52,13 @@ export ENCLAVE_PKG_NAME=erc20
 # make with backup disabled
 make DEBUG=1 ENCLAVE_DIR=example/erc20/enclave FEATURE_FLAGS="runtime_enabled,treekem"
 cd ${ANONIFY_ROOT}/tests/integration
-RUST_BACKTRACE=1 RUST_LOG=debug cargo test --no-default-features --features treekem -- --nocapture
+RUST_BACKTRACE=1 RUST_LOG=debug cargo test test_treekem --no-default-features --features treekem -- --nocapture
 
 cd ${ANONIFY_ROOT}/scripts
 
 make DEBUG=1 ENCLAVE_DIR=example/erc20/enclave FEATURE_FLAGS="runtime_enabled,enclave_key"
 cd ${ANONIFY_ROOT}/tests/integration
-RUST_BACKTRACE=1 RUST_LOG=debug cargo test -- --nocapture
+RUST_BACKTRACE=1 RUST_LOG=debug cargo test test_enclave_key -- --nocapture
 
 # Deploy a FACTORY Contract
 cd ${ANONIFY_ROOT}/anonify-contracts/deployer
