+++ tests/integration/Cargo.toml
@@ -22,8 +22,3 @@ ethabi = "12.0.0"
 serde_json = "1.0"
 rand = "0.7"
 once_cell = "1.5.2"
-
-[features]
-default = ["enclave_key"]
-enclave_key = []
-treekem = []
